# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ErrorDetails": {
				"anyOf": [
					{
						"properties": {
							"folderName": {
								"type": "string"
							},
							"userId": {
								"type": "string"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageUrl": {
								"type": "string"
							},
							"imageId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"userId": {
								"type": "string"
							},
							"challengeId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"longitude": {
								"type": "number",
								"format": "double"
							},
							"latitude": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"reason": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"searchKeyword": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/FieldErrors"
					},
					{
						"properties": {
							"extension": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageId": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				],
				"nullable": true
			},
			"ITsoaErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ErrorDetails"
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseFromMostTagToClient": {
				"properties": {
					"_count": {
						"properties": {
							"_all": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"_all"
						],
						"type": "object"
					},
					"tagCategoryId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"_count",
					"tagCategoryId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseFromMostTagToClient-Array_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"items": {
							"$ref": "#/components/schemas/ResponseFromMostTagToClient"
						},
						"type": "array"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseFromAward": {
				"properties": {
					"id": {
						"type": "string"
					},
					"awardMonth": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"awardMonth",
					"createdAt",
					"updatedAt",
					"status",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseFromAward_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ResponseFromAward"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseFromUpdateAward": {
				"properties": {
					"imageId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"awardId": {
						"type": "string"
					}
				},
				"required": [
					"imageId",
					"createdAt",
					"updatedAt",
					"status",
					"awardId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseFromUpdateAward-Array_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"items": {
							"$ref": "#/components/schemas/ResponseFromUpdateAward"
						},
						"type": "array"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseFromAward-Array_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"items": {
							"$ref": "#/components/schemas/ResponseFromAward"
						},
						"type": "array"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"sessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			}
		}
	},
	"info": {
		"title": "sweepic-server",
		"version": "1.0.0",
		"description": "Sweepic 프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Sweepic local server"
		},
		{
			"url": "http://3.37.137.212:3000",
			"description": "Sweepic server"
		}
	],
	"paths": {
...
		"/user/history/most_tagged/get": {
			"get": {
				"operationId": "GetMostTagged",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseFromMostTagToClient-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SRH-400",
												"reason": "유저 정보가 없습니다. 다시 로그인 해주세요.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-404",
												"reason": "조회를 요청한 데이터가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자의 사진들의 카테고리별로 가장 많은 태그를 조회합니다.",
				"summary": "인기 태그 조회 API",
				"tags": [
					"History"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/history/award/create": {
			"post": {
				"operationId": "CreateNewAward",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseFromAward_"
								}
							}
						}
					},
					"400": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SRH-400",
												"reason": "유저 정보가 없습니다. 다시 로그인 해주세요.",
												"data": null
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-400",
												"reason": "이미 해당 월의 어워드가 존재합니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-404",
												"reason": "조회를 요청한 데이터가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자의 어워드를 생성합니다.\r\n같은 달에 어워드가 이미 존재하는 경우 생성되지 않습니다.",
				"summary": "어워드 생성 API",
				"tags": [
					"Award"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/history/award/modify": {
			"patch": {
				"operationId": "ModifyAward",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseFromUpdateAward-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SRH-400",
												"reason": "유저 정보가 없습니다. 다시 로그인 해주세요.",
												"data": null
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-400",
												"reason": "어워드 사진 형식이 잘못되었습니다.",
												"data": null
											},
											"success": null
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-400",
												"reason": "어워드 업데이트를 실패했습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-404",
												"reason": "조회를 요청한 데이터가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "유저의 어워드에 사진을 등록합니다.\r\n총 5장의 사진만 등록 가능하며 원래 있던 사진들은 삭제하고 다시 등록합니다.\r\n배열에 이미지의 mediaId를 입력합니다.",
				"summary": "어워드 수정 API",
				"tags": [
					"Award"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "awardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "이미지 id의 배열",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "이미지 id의 배열"
							}
						}
					}
				}
			}
		},
		"/user/history/award/get": {
			"get": {
				"operationId": "GetAward",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseFromAward-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SRH-400",
												"reason": "유저 정보가 없습니다. 다시 로그인 해주세요.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "HIS-404",
												"reason": "조회를 요청한 데이터가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자의 어워드의 목록을 출력합니다.",
				"summary": "어워드 조회 API",
				"tags": [
					"/award"
				],
				"security": [],
				"parameters": []
			}
		},
...
	}
}